+++ bot/handlers/stats_handler.py
@@ -49,11 +49,11 @@ async def cmd_stats(update: Update, ctx: ContextTypes.DEFAULT_TYPE) -> None:
     await update.message.reply_text("\n".join(lines), parse_mode="Markdown")
 
 
-def _human_size(n: int) -> str:
+def _human_size(n: float) -> str:
     for unit in ("B", "KB", "MB", "GB"):
         if n < 1024:
             return f"{n:.1f} {unit}"
-        n /= 1024  # type: ignore[assignment]
+        n /= 1024
     return f"{n:.1f} TB"
